# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "SandboxCreated": {
        "description": "Returned exactly once, at provisioning -- the only time the plaintext API key is\never available, same rule as EndpointCreated.secret.",
        "properties": {
          "api_key": {
            "title": "Api Key",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "quotas": {
            "$ref": "#/components/schemas/SandboxQuotasRead"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "api_key",
          "expires_at",
          "quotas"
        ],
        "title": "SandboxCreated",
        "type": "object"
      },
      "SandboxQuotasRead": {
        "properties": {
          "max_endpoints": {
            "title": "Max Endpoints",
            "type": "integer"
          },
          "max_events": {
            "title": "Max Events",
            "type": "integer"
          },
          "rate_limit_requests_per_second": {
            "title": "Rate Limit Requests Per Second",
            "type": "number"
          },
          "ttl_minutes": {
            "title": "Ttl Minutes",
            "type": "integer"
          }
        },
        "required": [
          "max_endpoints",
          "max_events",
          "rate_limit_requests_per_second",
          "ttl_minutes"
        ],
        "title": "SandboxQuotasRead",
        "type": "object"
      },
      "SignatureVerifyRequest": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "description": "The raw request body exactly as sent, as a string.",
            "title": "Body",
            "type": "string"
          },
          "secret": {
            "title": "Secret",
            "type": "string"
          },
          "signature": {
            "title": "Signature",
            "type": "string"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "integer"
          },
          "tolerance_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Defaults to Settings.signature_tolerance_seconds.",
            "title": "Tolerance Seconds"
          }
        },
        "required": [
          "secret",
          "timestamp",
          "body",
          "signature"
        ],
        "title": "SignatureVerifyRequest",
        "type": "object"
      },
      "SignatureVerifyResponse": {
        "properties": {
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid"
        ],
        "title": "SignatureVerifyResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "description": "A Relay API key, in the form `<prefix>.<secret>`. Scoped to a single tenant; every request is authorized and tenant-scoped by this key.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Relay",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/sandbox": {
      "post": {
        "description": "No account needed: creates a scoped, TTL-limited (60 min) tenant and API key with hard-capped quotas (max endpoints, max events, a tight per-second rate limit), for driving the public demo console. Rate-limited per client IP.",
        "operationId": "create_sandbox_v1_sandbox_post",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "rate limit exceeded, retry after 1.00s",
                  "instance": "/v1/events",
                  "status": 429,
                  "title": "Too Many Requests",
                  "trace_id": "b3f1c2a4-5e6d-4f7a-8b9c-0d1e2f3a4b5c",
                  "type": "/problems/rate-limited"
                }
              }
            },
            "description": "The tenant's rate-limit budget is exhausted; retry after the value (seconds) in the Retry-After response header."
          }
        },
        "summary": "Provision a sandbox tenant and API key",
        "tags": [
          "sandbox"
        ]
      }
    },
...
    "/v1/sandbox/stream": {
      "get": {
        "description": "Server-Sent Events stream of the authenticated tenant's delivery attempts as they happen -- attempt #, status, latency, error class, next_retry_at, and breaker state. Scoped by construction: this subscribes to a per-tenant Redis Pub/Sub channel, so another tenant's events are never received, not merely filtered out.",
        "operationId": "stream_attempts_v1_sandbox_stream_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "missing or invalid API key",
                  "instance": "/v1/endpoints",
                  "status": 401,
                  "title": "Unauthorized",
                  "trace_id": "b3f1c2a4-5e6d-4f7a-8b9c-0d1e2f3a4b5c",
                  "type": "/problems/unauthorized"
                }
              }
            },
            "description": "Missing, malformed, unknown, or revoked API key."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "insufficient scope",
                  "instance": "/v1/endpoints",
                  "status": 403,
                  "title": "Forbidden",
                  "trace_id": "b3f1c2a4-5e6d-4f7a-8b9c-0d1e2f3a4b5c",
                  "type": "/problems/forbidden"
                }
              }
            },
            "description": "A valid key that lacks the scope this endpoint requires, or a sandbox tenant that has exceeded one of its hard-capped quotas."
          }
        },
        "summary": "Live delivery-attempt timeline (SSE)",
        "tags": [
          "sandbox"
        ]
      }
    },
    "/v1/sandbox/verify-signature": {
      "post": {
        "description": "Thin wrapper over relay.infra.signing.verify() -- the exact function every receiver should run on their end -- so the console's signature-verification widget exercises the real implementation rather than a JS reimplementation of it.",
        "operationId": "verify_signature_v1_sandbox_verify_signature_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignatureVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignatureVerifyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "missing or invalid API key",
                  "instance": "/v1/endpoints",
                  "status": 401,
                  "title": "Unauthorized",
                  "trace_id": "b3f1c2a4-5e6d-4f7a-8b9c-0d1e2f3a4b5c",
                  "type": "/problems/unauthorized"
                }
              }
            },
            "description": "Missing, malformed, unknown, or revoked API key."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "insufficient scope",
                  "instance": "/v1/endpoints",
                  "status": 403,
                  "title": "Forbidden",
                  "trace_id": "b3f1c2a4-5e6d-4f7a-8b9c-0d1e2f3a4b5c",
                  "type": "/problems/forbidden"
                }
              }
            },
            "description": "A valid key that lacks the scope this endpoint requires, or a sandbox tenant that has exceeded one of its hard-capped quotas."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "subscribed_event_types"
                      ],
                      "msg": "Field required",
                      "type": "missing"
                    }
                  ],
                  "instance": "/v1/endpoints",
                  "status": 422,
                  "title": "Validation Error",
                  "trace_id": "b3f1c2a4-5e6d-4f7a-8b9c-0d1e2f3a4b5c",
                  "type": "/problems/validation-error"
                }
              }
            },
            "description": "Request body/query failed validation, or a cursor was malformed."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Verify an HMAC delivery signature",
        "tags": [
          "sandbox"
        ]
      }
    }
  }
}